+++ zerver/actions/message_summary.py
@@ -15,6 +15,7 @@ from zerver.lib.narrow import (
     fetch_messages,
 )
 from zerver.models import UserProfile
+from zerver.models.realms import MessageEditHistoryVisibilityPolicyEnum
 
 # Maximum number of messages that can be summarized in a single request.
 MAX_MESSAGES_SUMMARIZED = 100
@@ -120,7 +121,7 @@ def do_summarize_narrow(
         client_gravatar=True,
         allow_empty_topic_name=False,
         # Avoid fetching edit history, which won't be passed to the model.
-        allow_edit_history=False,
+        message_edit_history_visibility_policy=MessageEditHistoryVisibilityPolicyEnum.none.value,
         user_profile=user_profile,
         realm=user_profile.realm,
     )