+++ zerver/tests/test_home.py
@@ -188,6 +188,7 @@ class HomeTest(ZulipTestCase):
         "realm_message_content_allowed_in_email_notifications",
         "realm_message_content_delete_limit_seconds",
         "realm_message_content_edit_limit_seconds",
+        "realm_message_edit_history_visibility_policy",
         "realm_message_retention_days",
         "realm_move_messages_between_streams_limit_seconds",
         "realm_move_messages_within_stream_limit_seconds",