+++ zerver/views/realm.py
@@ -1,4 +1,5 @@
 from collections.abc import Mapping
+from enum import Enum
 from typing import Annotated, Any, Literal
 
 from django.conf import settings
@@ -48,7 +49,11 @@ from zerver.lib.user_groups import (
 from zerver.lib.validator import check_capped_url, check_string
 from zerver.models import Realm, RealmReactivationStatus, RealmUserDefault, UserProfile
 from zerver.models.groups import SystemGroups
-from zerver.models.realms import DigestWeekdayEnum, OrgTypeEnum
+from zerver.models.realms import (
+    DigestWeekdayEnum,
+    MessageEditHistoryVisibilityPolicyEnum,
+    OrgTypeEnum,
+)
 from zerver.views.user_settings import (
     check_information_density_setting_values,
     check_settings_values,
@@ -112,7 +117,11 @@ def update_realm(
     message_content_edit_limit_seconds_raw: Annotated[
         Json[int | str] | None, ApiParamConfig("message_content_edit_limit_seconds")
     ] = None,
-    allow_edit_history: Json[bool] | None = None,
+    message_edit_history_visibility_policy_raw: Annotated[
+        # TODO: Use MessageEditHistoryVisibilityPolicyEnum here with Pydantic
+        Literal["all", "moves", "none"] | None,
+        ApiParamConfig("message_edit_history_visibility_policy"),
+    ] = None,
     default_language: str | None = None,
     waiting_period_threshold: Json[NonNegativeInt] | None = None,
     authentication_methods: Json[dict[str, Any]] | None = None,
@@ -328,6 +337,12 @@ def update_realm(
             )
 
             data["jitsi_server_url"] = jitsi_server_url
+
+    message_edit_history_visibility_policy: Enum | None = None
+    if message_edit_history_visibility_policy_raw is not None:
+        message_edit_history_visibility_policy = MessageEditHistoryVisibilityPolicyEnum[
+            message_edit_history_visibility_policy_raw
+        ]
 
     # The user of `locals()` here is a bit of a code smell, but it's
     # restricted to the elements present in realm.property_types.